+tons of cleanup. +compiles with mingwstudio now +no more extra winsock dependencies- created my own htons() and htonl() functions :-)
git-svn-id: http://miranda.googlecode.com/svn/branches/0_5_x@3253 83df23c2-51fd-11de-bb24-2d5e9b4af41d

+++ miranda/protocols/AimOscar/direct_connect.cpp
@@ -34,7 +34,7 @@ void aim_direct_connection_initiated(HANDLE hNewConnection, DWORD dwRemoteIP, vo
 	if(hContact)
 	{
 		ProtoBroadcastAck(AIM_PROTOCOL_NAME, hContact, ACKTYPE_FILE, ACKRESULT_CONNECTED,hContact, 0);
-		file_transfer_type=DBGetContactSettingByte(hContact,AIM_PROTOCOL_NAME,AIM_KEY_FT,-1);//okay now we see if they belong
+		file_transfer_type=(short)DBGetContactSettingByte(hContact,AIM_PROTOCOL_NAME,AIM_KEY_FT,-1);//okay now we see if they belong
 	}
 	if(file_transfer_type==1)//we are sending
 	{
